# core-banking-ledger-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-banking-ledger
  description: Banking Ledger Core Application
  contact:
    name: Firefly Software Solutions Inc
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.catalis.vc/ledger
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/transactions/{transactionId}/line-card:
    get:
      tags:
        - Transaction Line Card
      summary: Get Card Line
      description: Retrieve the card line record associated with the specified transaction.
      operationId: getCardLine
      parameters:
        - name: transactionId
          in: path
          description: Unique identifier of the transaction
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved the card line
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionLineCardDTO'
        '404':
          description: Card line not found for this transaction
    put:
      tags:
        - Transaction Line Card
      summary: Update Card Line
      description: Update an existing card line record associated with the specified transaction.
      operationId: updateCardLine
      parameters:
        - name: transactionId
          in: path
          description: Unique identifier of the transaction
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionLineCardDTO'
        required: true
      responses:
        '200':
          description: Card line updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionLineCardDTO'
        '404':
          description: Card line not found for this transaction
    post:
      tags:
        - Transaction Line Card
      summary: Create Card Line
      description: Create a new card line record associated with the specified transaction.
      operationId: createCardLine
      parameters:
        - name: transactionId
          in: path
          description: Unique identifier of the transaction
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionLineCardDTO'
        required: true
      responses:
        '201':
          description: Card line created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionLineCardDTO'
        '400':
          description: Invalid card line data provided
    delete:
      tags:
        - Transaction Line Card
      summary: Delete Card Line
      description: Remove an existing card line record from a specific transaction.
      operationId: deleteCardLine
      parameters:
        - name: transactionId
          in: path
          description: Unique identifier of the transaction
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Card line record deleted successfully
        '404':
          description: Card line record not found for this transaction
...
components:
  schemas:
...
    TransactionLineCardDTO:
      required:
        - transactionId
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        transactionLineCardId:
          type: string
          format: uuid
          readOnly: true
        transactionId:
          type: string
          format: uuid
        cardAuthCode:
          maxLength: 20
          minLength: 0
          type: string
        cardMerchantCategoryCode:
          maxLength: 10
          minLength: 0
          type: string
        cardMerchantName:
          maxLength: 200
          minLength: 0
          type: string
        cardPosEntryMode:
          maxLength: 10
          minLength: 0
          type: string
        cardTransactionReference:
          maxLength: 100
          minLength: 0
          type: string
        cardTerminalId:
          maxLength: 20
          minLength: 0
          type: string
        cardHolderCountry:
          maxLength: 3
          minLength: 0
          type: string
        cardPresentFlag:
          type: boolean
        cardTransactionTimestamp:
          type: string
          format: date-time
        cardFraudFlag:
          type: boolean
        cardCurrencyConversionRate:
          minimum: 0
          exclusiveMinimum: false
          type: number
        cardFeeAmount:
          type: number
        cardFeeCurrency:
          type: string
        cardInstallmentPlan:
          maxLength: 50
          minLength: 0
          type: string
        cardMerchantCif:
          maxLength: 20
          minLength: 0
          type: string
...